refactor(embedded-checkout): CHECKOUT-3660 Create no-op EmbeddedCheckoutMessenger if not inside iframe
