+++ src/embedded-checkout/iframe-content/create-embedded-checkout-messenger.spec.ts
@@ -1,9 +1,15 @@
 import createEmbeddedCheckoutMessenger from './create-embedded-checkout-messenger';
-import EmbeddedCheckoutMessenger from './embedded-checkout-messenger';
+import IframeEmbeddedCheckoutMessenger from './iframe-embedded-checkout-messenger';
+import NoopEmbeddedCheckoutMessenger from './noop-embedded-checkout-messenger';
 
 describe('createEmbeddedCheckoutMessenger()', () => {
-    it('creates embedded checkout messenger', () => {
+    it('creates no-op embedded checkout messenger if not inside iframe', () => {
         expect(createEmbeddedCheckoutMessenger({ parentOrigin: 'https://foobar.mybigcommerece.com' }))
-            .toBeInstanceOf(EmbeddedCheckoutMessenger);
+            .toBeInstanceOf(NoopEmbeddedCheckoutMessenger);
+    });
+
+    it('creates embedded checkout messenger if inside iframe', () => {
+        expect(createEmbeddedCheckoutMessenger({ parentOrigin: 'https://foobar.mybigcommerece.com', parentWindow: Object.create(window) }))
+            .toBeInstanceOf(IframeEmbeddedCheckoutMessenger);
     });
 });
