+++ src/embedded-checkout/iframe-content/embedded-checkout-messenger-options.ts
@@ -1,3 +1,4 @@
 export default interface EmbeddedCheckoutMessengerOptions {
     parentOrigin: string;
+    parentWindow?: Window;
 }
